Add abort functionality and reset button to the sorting algorithms
Adding abort functionality and a reset button to the sorting algorithms to allow the user to cancel an ongoing sorting operation and reset the array. The code now utilizes the AbortController and AbortSignal objects to abort the ongoing sorting operation and reset the array accordingly. This ensures that the sorting algorithms can be terminated prematurely if desired by the user.